pyproject.toml
[tool.poetry]
name = "hexlet_code"
version = "0.2.0"
description = ""
authors = ["mistermikebell <jokerbell7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
page_loader = "page_loader.scripts.page_loader:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
